+++ sshd-core/src/main/java/org/apache/sshd/common/kex/SNTRUP761.java
@@ -21,7 +21,6 @@ package org.apache.sshd.common.kex;
 import java.security.SecureRandom;
 import java.util.Arrays;
 
-import org.apache.sshd.common.util.security.SecurityUtils;
 import org.bouncycastle.crypto.AsymmetricCipherKeyPair;
 import org.bouncycastle.crypto.SecretWithEncapsulation;
 import org.bouncycastle.pqc.crypto.ntruprime.SNTRUPrimeKEMExtractor;
@@ -42,9 +41,6 @@ final class SNTRUP761 {
     }
 
     static boolean isSupported() {
-        if (!SecurityUtils.isBouncyCastleRegistered()) {
-            return false;
-        }
         try {
             return SNTRUPrimeParameters.sntrup761.getSessionKeySize() == 256; // BC < 1.78 had only 128
         } catch (Throwable e) {